Changes to db tables and insert script

+++ db/objects/intervention_costings/V0.1.0.201__food_vehicle.sql
@@ -1,6 +1,6 @@
 
 CREATE TABLE food_vehicle(
-    id              numeric   PRIMARY KEY ,
+    id              integer   PRIMARY KEY ,
     vehicle_name    text
 )
 ;
